Remove repetitive code for switch cpu cores and package this function to a generic function.

+++ Sources/AMDVTDriver/AMDVTDriver/SVM.h
@@ -192,7 +192,7 @@ enum SVMStatus
 class SVMManager : public IManager
 {
 	VirtCpuInfo** pVirtCpuInfo;
-	UINT64 cpuCnt;
+	UINT32 cpuCnt;
 	MsrPremissionsMapManager msrPremissionMap;
 	IMsrInterceptPlugin* pMsrInterceptPlugin;
 	ICpuidInterceptPlugin* pCpuIdInterceptPlugin;
